feat: new file to contain ArrayDepth struct
